refactor: Add paths alias, Remove relative paths
