+++ src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import '../styles/global.css'
+import '@/styles/global.css'
 import 'katex/dist/katex.css'
 import Footer from '../components/footer'
