+++ src/pages/blog/[slug].tsx
@@ -1,18 +1,18 @@
 import Link from 'next/link'
 import fetch from 'node-fetch'
 import { useRouter } from 'next/router'
-import Header from '../../components/header'
-import Heading from '../../components/heading'
-import components from '../../components/dynamic'
+import Header from '@/components/header'
+import Heading from '@/components/heading'
+import components from '@/components/dynamic'
 import ReactJSXParser from '@zeit/react-jsx-parser'
-import blogStyles from '../../styles/blog.module.css'
-import { textBlock } from '../../lib/notion/renderers'
-import { values } from '../../lib/notion/rpc'
-import getPageData from '../../lib/notion/getPageData'
+import blogStyles from '@/styles/blog.module.css'
+import { textBlock } from '@/lib/notion/renderers'
+import { values } from '@/lib/notion/rpc'
+import getPageData from '@/lib/notion/getPageData'
 import React, { CSSProperties, useEffect } from 'react'
-import getBlogIndex from '../../lib/notion/getBlogIndex'
-import getNotionUsers from '../../lib/notion/getNotionUsers'
-import { getBlogLink, getDateStr } from '../../lib/blog-helpers'
+import getBlogIndex from '@/lib/notion/getBlogIndex'
+import getNotionUsers from '@/lib/notion/getNotionUsers'
+import { getBlogLink, getDateStr } from '@/lib/blog-helpers'
 
 // Get the data for each blog post
 export async function getStaticProps({ params: { slug }, preview }) {
