+++ src/pages/contact.tsx
@@ -1,12 +1,12 @@
-import Header from '../components/header'
-import ExtLink from '../components/ext-link'
+import Header from '@/components/header'
+import ExtLink from '@/components/ext-link'
 
-import sharedStyles from '../styles/shared.module.css'
-import contactStyles from '../styles/contact.module.css'
+import sharedStyles from '@/styles/shared.module.css'
+import contactStyles from '@/styles/contact.module.css'
 
-import GitHub from '../components/svgs/github'
-import Envelope from '../components/svgs/envelope'
-import LinkedIn from '../components/svgs/linkedin'
+import GitHub from '@/components/svgs/github'
+import Envelope from '@/components/svgs/envelope'
+import LinkedIn from '@/components/svgs/linkedin'
 
 const contacts = [
   {
