+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
-import Header from '../components/header'
-import ExtLink from '../components/ext-link'
-import Features from '../components/features'
-import sharedStyles from '../styles/shared.module.css'
+import Header from '@/components/header'
+import ExtLink from '@/components/ext-link'
+import Features from '@/components/features'
+import sharedStyles from '@/styles/shared.module.css'
 
 export default function Index() {
   return (
