+++ src/styles/header.module.css
@@ -23,7 +23,7 @@
 
 .homeLogo {
   font-size: 2rem;
-  width: 2rem;
+  width: 6rem;
   cursor: pointer;
 }
